requirements.txt
dramatiq[rabbitmq, watch]~=1.16.0
fastapi~=0.110.1
pydantic[dotenv]~=2.6.4
pydantic-settings
python-multipart
psycopg2-binary
sqlmodel~=0.0.16
uvicorn[standard]
pypdfium2~=4.28.0
pillow~=10.3.0
aiofiles~=23.2.1
types-aiofiles

# dev
pytest
httpx

